try to handle using multiple packages better
Globals are defined in environment corresponding to SYMBOL-PACKAGE of
the name, and packages can be locked to prevent accidental
redefinitions of shared symbols.

CL and 3BGL-GLSL packages are locked, but CL:POSITION vertex attribute
is predefined as '(input position :vec4 :location 0)' so user code can
use it. (redefinition with same definition is allowed)